+++ OneBlink.SDK/models/AttachmentUploadCredentialsRequest.cs
@@ -1,0 +1,7 @@
+namespace OneBlink.SDK.Model
+{
+    public class AttachmentUploadCredentialsRequest
+    {
+        public string username {get;set;}
+    }
+}