+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `storageType` to `FormElement`
 - `formsAppEnvironment` to `FormsClient Search`
 - `GetFormSubmissionAttachment` to `FormsClient`
+- `CreateSubmissionAttachment` to `FormsClient`
 
 ## 2.1.0 (2021-04-27)